Predicting Kentucky's Failing Schools
Kentucky ranks 45th in the nation in uneducated population-- only 5 states have a worse education system. The average state dropout rate is 6.5%. The purpose of this research is to provide the state of Kentucky with some guidelines moving forward in order to reduce drop out rates. First by building a model to predict dropout rates then by using these models to determine which features help or hinder the most.